+++ ros/drake_ros_systems/include/drake/ros/systems/ros_sensor_publisher_lidar.h
@@ -17,6 +17,7 @@ using drake::RigidBodyDepthSensor;
 
 namespace drake {
 namespace ros {
+namespace systems {
 
 /**
  * @brief A system that takes the system state as the input,
@@ -271,5 +272,6 @@ class SensorPublisherLidar {
   ::ros::Time previous_send_time_;
 };
 
-}  // end namespace ros
-}  // end namespace drake
+}  // namespace systems
+}  // namespace ros
+}  // namespace drake
